+++ modules/store-review/src/main/java/com/gluonhq/attach/storereview/impl/IOSStoreReviewService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2019 Gluon
+ * Copyright (c) 2021, Gluon
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -25,9 +25,23 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package com.gluonhq.attach.storereview.impl;
 
-/**
- * Primary API package for Attach - Review plugin,
- * contains the interface {@link com.gluonhq.attach.storereview.StoreReviewService} and related classes.
- */
-package com.gluonhq.attach.storereview;
+import com.gluonhq.attach.storereview.StoreReviewService;
+import javafx.application.Platform;
+
+public class IOSStoreReviewService implements StoreReviewService {
+
+    static {
+       Platform.runLater(() -> System.loadLibrary("StoreReview"));
+    }
+    
+    @Override
+    public void requestStoreReview() {
+        nativeRequestStoreReview();
+    }
+
+    // native
+    private static native void nativeRequestStoreReview();
+
+}
