+++ modules/store-review/src/main/native/android/c/storereview.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Gluon
+ * Copyright (c) 2021, Gluon
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
