+++ modules/store-review/src/main/native/ios/StoreReview.m
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, Gluon
+ * Copyright (c) 2021, Gluon
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -25,23 +25,32 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.attach.storereview.impl;
+#import <StoreKit/StoreKit.h>
+#include "jni.h"
 
-import com.gluonhq.attach.storereview.StoreReviewService;
-import javafx.application.Platform;
+JNIEnv *env;
 
-public class IOSStoreReviewService implements StoreReviewService {
-
-    static {
-       Platform.runLater(() -> System.loadLibrary("StoreReview"));
+JNIEXPORT jint JNICALL
+JNI_OnLoad_StoreReview(JavaVM *vm, void *reserved)
+{
+#ifdef JNI_VERSION_1_8
+    //min. returned JNI_VERSION required by JDK8 for builtin libraries
+    if ((*vm)->GetEnv(vm, (void **)&env, JNI_VERSION_1_8) != JNI_OK) {
+        return JNI_VERSION_1_4;
     }
-    
-    @Override
-    public void requestStoreReview() {
-        nativeRequestStoreReview();
+    return JNI_VERSION_1_8;
+#else
+    return JNI_VERSION_1_4;
+#endif
+}
+
+
+JNIEXPORT void JNICALL Java_com_gluonhq_attach_storereview_impl_IOSStoreReviewService_nativeRequestStoreReview
+(JNIEnv *env, jclass jClass) {
+    if (@available(iOS 10.3, *)) {
+        if (debugAttach) {
+            AttachLog(@"Calling requestReview:");
+        }
+        [SKStoreReviewController requestReview];
     }
-
-    // native
-    private static native void nativeRequestStoreReview();
-
 }
