+++ modules/store-review/src/main/resources/META-INF/substrate/config/reflectionconfig-aarch64-android.json
